+++ ModernCppProject/ModernCppProject/GetAnswerHandler.h
@@ -7,11 +7,11 @@
 class GetAnswerHandler
 {
 public:
-	GetAnswerHandler(std::vector<std::tuple<uint8_t, float, float>> &answers);
+	GetAnswerHandler(std::vector<std::tuple<int, float, float>> &answers);
 
 	/*Operator overload*/
 	void operator() (const crow::request& req, crow::response& res, int userId) const;
 private:
-	std::vector<std::tuple<uint8_t, float, float>>* m_answers;
+	std::vector<std::tuple<int, float, float>>* m_answers;
 };
